examples/transforming/pyproject.toml
[build-system]
requires = ['setuptools']
build-backend = 'setuptools.build_meta'

[project]
name = 'metadata_transformers'
description = 'Metadata propagation transformers'
version = '0.1.1'
readme = 'README.md'
requires-python = '>=3.9'
dependencies = [
  'acryl-datahub[dbt]',
  'click',
  'jmespath',
  'datahub-tools@git+ssh://git@github.com/makenotion/datahub-tools',
]

[tool.setuptools]
include-package-data = false
package-dir = { "" = "src" }

[tool.isort]
profile = 'black'
line_length = 88

[tool.black]
target-version = ['py310']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
